Revert rename for `impl.expr`. Rename `asserts.segment.expression` to `asserts.segment.expr`.
